[refactoring] 리뷰 반영하여 Car 객체에서 이름 체크하도록 수정
